+++ app/src/main/java/org/ole/planet/myplanet/lite/DashboardVoicesFragment.kt
@@ -8,54 +8,37 @@ package org.ole.planet.myplanet.lite
 
 import android.app.Activity
 import android.content.Context
-import android.content.Intent
-import android.icu.text.CompactDecimalFormat
 import android.net.Uri
 import android.os.Bundle
-import android.text.TextUtils
 import android.view.View
-import android.view.ViewGroup
-import android.widget.ImageView
-import android.widget.LinearLayout
 import android.widget.TextView
 import android.widget.Toast
 import androidx.activity.result.contract.ActivityResultContracts
-import androidx.appcompat.widget.AppCompatImageView
 import androidx.core.view.isVisible
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.lifecycleScope
-import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.google.android.material.floatingactionbutton.FloatingActionButton
-import com.squareup.moshi.Moshi
-import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
 import io.noties.markwon.Markwon
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
-import okhttp3.OkHttpClient
 import org.ole.planet.myplanet.lite.auth.AuthDependencies
-import org.ole.planet.myplanet.lite.dashboard.CreateVoiceActivity
 import org.ole.planet.myplanet.lite.dashboard.DashboardAvatarLoader
-import org.ole.planet.myplanet.lite.dashboard.DashboardImagePreviewActivity
 import org.ole.planet.myplanet.lite.dashboard.DashboardNewsActionsRepository
 import org.ole.planet.myplanet.lite.dashboard.DashboardNewsRepository
-import org.ole.planet.myplanet.lite.dashboard.SharedBitmapDependencies
-import org.ole.planet.myplanet.lite.dashboard.DashboardNewsRepository.NewsDocument
 import org.ole.planet.myplanet.lite.dashboard.DashboardNewsRepository.NewsPage
 import org.ole.planet.myplanet.lite.dashboard.DashboardPostDetailActivity
 import org.ole.planet.myplanet.lite.dashboard.DashboardPostImageLoader
 import org.ole.planet.myplanet.lite.dashboard.DashboardServerPreferences
-import org.ole.planet.myplanet.lite.dashboard.DashboardTeamMemberProfileActivity
 import org.ole.planet.myplanet.lite.dashboard.PostShareHelper
+import org.ole.planet.myplanet.lite.dashboard.SharedBitmapDependencies
 import org.ole.planet.myplanet.lite.profile.AvatarUpdateNotifier
 import org.ole.planet.myplanet.lite.profile.ProfileCredentialsStore
 import org.ole.planet.myplanet.lite.profile.StoredCredentials
 import org.ole.planet.myplanet.lite.profile.UserProfile
 import org.ole.planet.myplanet.lite.profile.UserProfileDatabase
 import org.ole.planet.myplanet.lite.util.enableDrag
-import java.util.ArrayList
-import java.util.Locale
 
 class DashboardVoicesFragment : Fragment(R.layout.fragment_dashboard_voices) {
     private lateinit var recyclerView: RecyclerView
@@ -98,6 +81,8 @@ class DashboardVoicesFragment : Fragment(R.layout.fragment_dashboard_voices) {
             moshi = AuthDependencies.moshi,
         )
     private val actionsRepository = DashboardNewsActionsRepository(AuthDependencies.client, AuthDependencies.moshi, Dispatchers.IO)
+    private val itemMapper by lazy { DashboardNewsItemMapper(requireContext()) }
+    private val navigator by lazy { DashboardVoicesNavigator(this, createVoiceLauncher, postDetailLauncher) }
 
     @androidx.annotation.VisibleForTesting
     internal val items = mutableListOf<DashboardNewsItem>()
@@ -156,72 +141,28 @@ class DashboardVoicesFragment : Fragment(R.layout.fragment_dashboard_voices) {
 
     private fun setupRecyclerView() {
         markwon = Markwon.builder(requireContext()).build()
-        adapter = createDashboardNewsAdapter()
-
-        recyclerView.layoutManager = LinearLayoutManager(requireContext())
-        recyclerView.adapter = adapter
-        setupScrollListener()
-    }
-
-    private fun createDashboardNewsAdapter(): DashboardNewsAdapter =
-        DashboardNewsAdapter(
-            markwon,
-            avatarBinder = { imageView, username, hasAvatar ->
-                avatarLoader?.bind(imageView, username, hasAvatar)
-            },
-            imageBinder = { imageView, imagePath ->
-                val loader = postImageLoader
-                if (loader != null) {
-                    loader.bind(imageView, imagePath)
-                } else {
-                    imageView.isVisible = false
-                    imageView.setImageDrawable(null)
-                }
-            },
-            onImageClicked = { item, index ->
-                openImagePreview(item, index)
-            },
-            onPostClicked = { item ->
-                openPostDetail(item)
-            },
-            onDeleteClicked = { item ->
-                requestDeletePost(item)
-            },
-            onShareClicked = { item ->
-                sharePost(item)
-            },
-            onEditClicked = { item ->
-                openEditVoice(item)
-            },
-            onAuthorClicked = { item ->
-                openTeamMemberProfile(item)
-            },
-        )
-
-    private fun setupScrollListener() {
-        recyclerView.addOnScrollListener(
-            object : RecyclerView.OnScrollListener() {
-                override fun onScrolled(
-                    recyclerView: RecyclerView,
-                    dx: Int,
-                    dy: Int,
-                ) {
-                    super.onScrolled(recyclerView, dx, dy)
-                    if (dy <= 0) {
-                        return
+        adapter =
+            DashboardVoicesRecyclerBinder.bind(
+                recyclerView,
+                markwon,
+                shouldLoadMore = { !isLoading && hasMore },
+                loadMore = { loadMore(pageSize) },
+                avatarBinder = { imageView, username, hasAvatar ->
+                    avatarLoader?.bind(imageView, username, hasAvatar)
+                },
+                imageBinder = { imageView, imagePath ->
+                    postImageLoader?.bind(imageView, imagePath) ?: run {
+                        imageView.isVisible = false
+                        imageView.setImageDrawable(null)
                     }
-                    val layoutManager = recyclerView.layoutManager as? LinearLayoutManager ?: return
-                    val visibleItemCount = layoutManager.childCount
-                    val totalItemCount = layoutManager.itemCount
-                    val firstVisibleItemPosition = layoutManager.findFirstVisibleItemPosition()
-                    if (!isLoading && hasMore && totalItemCount > 0) {
-                        if (firstVisibleItemPosition + visibleItemCount >= totalItemCount - LOAD_MORE_THRESHOLD) {
-                            loadMore(pageSize)
-                        }
-                    }
-                }
-            },
-        )
+                },
+                onImageClicked = ::openImagePreview,
+                onPostClicked = ::openPostDetail,
+                onDeleteClicked = ::requestDeletePost,
+                onShareClicked = ::sharePost,
+                onEditClicked = ::openEditVoice,
+                onAuthorClicked = ::openTeamMemberProfile,
+            )
     }
 
     private fun setupObserversAndLoadInitial() {
@@ -282,30 +223,7 @@ class DashboardVoicesFragment : Fragment(R.layout.fragment_dashboard_voices) {
     }
 
     private fun handleAvatarUpdated(username: String) {
-        if (!::adapter.isInitialized) {
-            return
-        }
-
-        val positions =
-            if (adapter.isIndexAvailable()) {
-                adapter.getPositionsForUsername(username)
-            } else {
-                adapter.currentList.mapIndexedNotNull { index, item ->
-                    if (item.username?.equals(username, ignoreCase = true) == true) {
-                        index
-                    } else {
-                        null
-                    }
-                }
-            }
-        if (positions.isEmpty()) {
-            return
-        }
-        recyclerView.post {
-            positions.forEach { position ->
-                adapter.notifyItemChanged(position)
-            }
-        }
+        if (::adapter.isInitialized) adapter.notifyAvatarUpdated(recyclerView, username)
     }
 
     private suspend fun initializeSession() {
@@ -421,7 +339,7 @@ class DashboardVoicesFragment : Fragment(R.layout.fragment_dashboard_voices) {
             page.items.mapNotNull { document ->
                 val id = document.id ?: return@mapNotNull null
                 val commentCount = commentCounts[id] ?: 0
-                mapToItem(document, commentCount)
+                itemMapper.map(document, commentCount, sessionCookie, currentUsername, isUserAdmin)
             }
         if (mapped.isNotEmpty()) {
             items.addAll(mapped)
@@ -511,11 +429,7 @@ class DashboardVoicesFragment : Fragment(R.layout.fragment_dashboard_voices) {
     }
 
     private fun openCreateVoiceComposer() {
-        val context = context ?: return
-        val intent = Intent(context, CreateVoiceActivity::class.java)
-        teamId?.let { intent.putExtra(CreateVoiceActivity.EXTRA_TARGET_TEAM_ID, it) }
-        teamName?.let { intent.putExtra(CreateVoiceActivity.EXTRA_TARGET_TEAM_NAME, it) }
-        createVoiceLauncher.launch(intent)
+        navigator.openComposer(teamId, teamName)
     }
 
     private fun showEmptyState(messageRes: Int) {
@@ -523,119 +437,6 @@ class DashboardVoicesFragment : Fragment(R.layout.fragment_dashboard_voices) {
         emptyView.setText(messageRes)
         emptyView.isVisible = true
     }
-
-    private fun mapToItem(
-        document: NewsDocument,
-        commentCount: Int,
-    ): DashboardNewsItem? {
-        val id = document.id ?: return null
-        val username = document.user?.name?.takeIf { it.isNotBlank() }
-        val displayName =
-            document.user?.let { user ->
-                val parts =
-                    listOfNotNull(user.firstName, user.middleName, user.lastName)
-                        .map { it.trim() }
-                        .filter { it.isNotEmpty() }
-                when {
-                    parts.isNotEmpty() -> TextUtils.join(" ", parts)
-                    !username.isNullOrEmpty() -> username
-                    else -> getString(R.string.dashboard_profile_name_placeholder)
-                }
-            } ?: (username ?: getString(R.string.dashboard_profile_name_placeholder))
-        val timeMillis = document.time ?: 0L
-        val relativeTime =
-            org.ole.planet.myplanet.lite.util.DateUtils
-                .formatRelativeTime(requireContext(), timeMillis)
-        val metadata =
-            org.ole.planet.myplanet.lite.util.DateUtils
-                .buildMetadata(username, relativeTime)
-        val rawMessage = document.message
-        val messageImages = extractImagePaths(rawMessage)
-        val documentImages = mapDocumentImages(document)
-        val imagePaths = (messageImages + documentImages).distinct()
-        val message = rawMessage?.trim()?.takeIf { it.isNotEmpty() }
-        val hasAvatar = document.user?.attachments?.containsKey("img") == true
-        val hasSession = !sessionCookie.isNullOrBlank()
-        val isAuthor = hasSession && !username.isNullOrBlank() && currentUsername?.equals(username, ignoreCase = true) == true
-        return DashboardNewsItem(
-            id = id,
-            author = displayName,
-            username = username,
-            metadata = metadata,
-            message = message,
-            hasAvatar = hasAvatar,
-            imagePaths = imagePaths,
-            commentCount = commentCount,
-            timestamp = timeMillis,
-            canEdit = isAuthor,
-            canDelete = isAuthor || isUserAdmin,
-            canShare = hasSession,
-            document = document,
-        )
-    }
-
-    private fun extractImagePaths(raw: String?): List<String> {
-        if (raw.isNullOrBlank()) {
-            return emptyList()
-        }
-        return IMAGE_MARKDOWN_CAPTURE_REGEX
-            .findAll(raw)
-            .mapNotNull { match ->
-                match.groupValues
-                    .getOrNull(1)
-                    ?.trim()
-                    .orEmpty()
-                    .takeIf { it.isNotBlank() }
-            }.toList()
-    }
-
-    private fun mapDocumentImages(document: NewsDocument): List<String> =
-        document.images
-            ?.mapNotNull { image ->
-                extractImagePath(image.markdown)
-                    ?: buildResourcePath(image.resourceId, image.filename)
-            }?.filter { it.isNotBlank() }
-            .orEmpty()
-
-    private fun extractImagePath(markdown: String?): String? {
-        if (markdown.isNullOrBlank()) {
-            return null
-        }
-        val match = IMAGE_MARKDOWN_CAPTURE_REGEX.find(markdown) ?: return null
-        return match.groupValues
-            .getOrNull(1)
-            ?.trim()
-            ?.takeIf { it.isNotEmpty() }
-    }
-
-    private fun buildResourcePath(
-        resourceId: String?,
-        filename: String?,
-    ): String? {
-        val id = resourceId?.trim().takeUnless { it.isNullOrEmpty() }
-        val name = filename?.trim().takeUnless { it.isNullOrEmpty() }
-        if (id == null || name == null) {
-            return null
-        }
-        return "resources/$id/$name"
-    }
-
-    @androidx.annotation.VisibleForTesting
-    internal data class DashboardNewsItem(
-        val id: String,
-        val author: String,
-        val username: String?,
-        val metadata: String,
-        val message: String?,
-        val hasAvatar: Boolean,
-        val imagePaths: List<String>,
-        val commentCount: Int,
-        val timestamp: Long,
-        val canEdit: Boolean,
-        val canDelete: Boolean,
-        val canShare: Boolean,
-        val document: NewsDocument,
-    )
 
     private suspend fun loadCurrentUserProfile(): UserProfile? {
         val context = requireContext().applicationContext
@@ -648,37 +449,11 @@ class DashboardVoicesFragment : Fragment(R.layout.fragment_dashboard_voices) {
         item: DashboardNewsItem,
         index: Int,
     ) {
-        val context = context ?: return
-        if (item.imagePaths.isEmpty()) {
-            return
-        }
-        val intent = Intent(context, DashboardImagePreviewActivity::class.java)
-        intent.putStringArrayListExtra(
-            DashboardImagePreviewActivity.EXTRA_IMAGE_PATHS,
-            ArrayList(item.imagePaths),
-        )
-        intent.putExtra(DashboardImagePreviewActivity.EXTRA_START_INDEX, index)
-        startActivity(intent)
+        navigator.openImagePreview(item, index)
     }
 
     private fun openPostDetail(item: DashboardNewsItem) {
-        val context = context ?: return
-        val intent = Intent(context, DashboardPostDetailActivity::class.java)
-        intent.putExtra(DashboardPostDetailActivity.EXTRA_POST_ID, item.id)
-        intent.putExtra(DashboardPostDetailActivity.EXTRA_AUTHOR, item.author)
-        intent.putExtra(DashboardPostDetailActivity.EXTRA_USERNAME, item.username)
-        intent.putExtra(DashboardPostDetailActivity.EXTRA_MESSAGE, item.message)
-        intent.putStringArrayListExtra(
-            DashboardPostDetailActivity.EXTRA_IMAGE_PATHS,
-            ArrayList(item.imagePaths),
-        )
-        intent.putExtra(DashboardPostDetailActivity.EXTRA_HAS_AVATAR, item.hasAvatar)
-        intent.putExtra(DashboardPostDetailActivity.EXTRA_TIMESTAMP, item.timestamp)
-        intent.putExtra(DashboardPostDetailActivity.EXTRA_COMMENT_COUNT, item.commentCount)
-        intent.putExtra(DashboardPostDetailActivity.EXTRA_DOCUMENT, item.document)
-        teamId?.let { intent.putExtra(DashboardPostDetailActivity.EXTRA_TEAM_ID, it) }
-        teamName?.let { intent.putExtra(DashboardPostDetailActivity.EXTRA_TEAM_NAME, it) }
-        postDetailLauncher.launch(intent)
+        navigator.openPostDetail(item, teamId, teamName)
     }
 
     private fun sharePost(item: DashboardNewsItem) {
@@ -689,41 +464,11 @@ class DashboardVoicesFragment : Fragment(R.layout.fragment_dashboard_voices) {
     }
 
     private fun openTeamMemberProfile(item: DashboardNewsItem) {
-        val username = item.username
-        if (username.isNullOrBlank()) {
-            Toast
-                .makeText(
-                    requireContext(),
-                    R.string.dashboard_team_members_profile_unavailable,
-                    Toast.LENGTH_SHORT,
-                ).show()
-            return
-        }
-        val displayName = item.author.ifBlank { username }
-        val intent =
-            DashboardTeamMemberProfileActivity.buildIntent(
-                requireContext(),
-                username,
-                displayName,
-                false,
-            )
-        startActivity(intent)
+        navigator.openTeamMemberProfile(item)
     }
 
     private fun openEditVoice(item: DashboardNewsItem) {
-        val context = context ?: return
-        val intent = Intent(context, CreateVoiceActivity::class.java)
-        intent.putExtra(CreateVoiceActivity.EXTRA_IS_EDIT_MODE, true)
-        intent.putExtra(CreateVoiceActivity.EXTRA_EDIT_POST_ID, item.id)
-        intent.putExtra(CreateVoiceActivity.EXTRA_EDIT_INITIAL_MESSAGE, item.message)
-        intent.putStringArrayListExtra(
-            CreateVoiceActivity.EXTRA_EDIT_INITIAL_IMAGE_PATHS,
-            ArrayList(item.imagePaths),
-        )
-        intent.putExtra(CreateVoiceActivity.EXTRA_EDIT_DOCUMENT, item.document)
-        teamId?.let { intent.putExtra(CreateVoiceActivity.EXTRA_TARGET_TEAM_ID, it) }
-        teamName?.let { intent.putExtra(CreateVoiceActivity.EXTRA_TARGET_TEAM_NAME, it) }
-        createVoiceLauncher.launch(intent)
+        navigator.openEditVoice(item, teamId, teamName)
     }
 
     fun isTeamFeedFor(
@@ -735,205 +480,9 @@ class DashboardVoicesFragment : Fragment(R.layout.fragment_dashboard_voices) {
         return sameId && sameName
     }
 
-    @androidx.annotation.VisibleForTesting
-    internal class DashboardNewsAdapter(
-        private val markwon: Markwon,
-        private val avatarBinder: (ImageView, String?, Boolean) -> Unit,
-        private val imageBinder: (ImageView, String) -> Unit,
-        private val onImageClicked: (DashboardNewsItem, Int) -> Unit,
-        private val onPostClicked: (DashboardNewsItem) -> Unit,
-        private val onDeleteClicked: (DashboardNewsItem) -> Unit,
-        private val onShareClicked: (DashboardNewsItem) -> Unit,
-        private val onEditClicked: (DashboardNewsItem) -> Unit,
-        private val onAuthorClicked: (DashboardNewsItem) -> Unit,
-    ) : androidx.recyclerview.widget.ListAdapter<DashboardNewsItem, DashboardNewsViewHolder>(DIFF_CALLBACK) {
-        override fun onCreateViewHolder(
-            parent: android.view.ViewGroup,
-            viewType: Int,
-        ): DashboardNewsViewHolder {
-            val view =
-                android.view.LayoutInflater
-                    .from(parent.context)
-                    .inflate(R.layout.item_dashboard_post, parent, false)
-            return DashboardNewsViewHolder(
-                view,
-                markwon,
-                avatarBinder,
-                imageBinder,
-                onImageClicked,
-                onPostClicked,
-                onDeleteClicked,
-                onShareClicked,
-                onEditClicked,
-                onAuthorClicked,
-            )
-        }
-
-        override fun onBindViewHolder(
-            holder: DashboardNewsViewHolder,
-            position: Int,
-        ) {
-            holder.bind(getItem(position))
-        }
-
-        private var usernameIndex: Map<String, List<Int>> = emptyMap()
-
-        override fun onCurrentListChanged(
-            previousList: List<DashboardNewsItem>,
-            currentList: List<DashboardNewsItem>,
-        ) {
-            super.onCurrentListChanged(previousList, currentList)
-            usernameIndex =
-                currentList
-                    .mapIndexedNotNull { index, item ->
-                        item.username?.let { it.lowercase() to index }
-                    }.groupBy({ it.first }, { it.second })
-        }
-
-        fun getPositionsForUsername(username: String): List<Int> = usernameIndex[username.lowercase()] ?: emptyList()
-
-        fun isIndexAvailable(): Boolean = usernameIndex.isNotEmpty()
-
-        companion object {
-            private val DIFF_CALLBACK =
-                org.ole.planet.myplanet.lite.util.DiffUtils.itemCallback<DashboardNewsItem>({ oldItem, newItem ->
-                    oldItem.id ==
-                        newItem.id
-                })
-        }
-    }
-
-    internal class DashboardNewsViewHolder(
-        view: View,
-        private val markwon: Markwon,
-        private val avatarBinder: (ImageView, String?, Boolean) -> Unit,
-        private val imageBinder: (ImageView, String) -> Unit,
-        private val onImageClicked: (DashboardNewsItem, Int) -> Unit,
-        private val onPostClicked: (DashboardNewsItem) -> Unit,
-        private val onDeleteClicked: (DashboardNewsItem) -> Unit,
-        private val onShareClicked: (DashboardNewsItem) -> Unit,
-        private val onEditClicked: (DashboardNewsItem) -> Unit,
-        private val onAuthorClicked: (DashboardNewsItem) -> Unit,
-    ) : RecyclerView.ViewHolder(view) {
-        private val authorView: TextView = view.findViewById(R.id.postAuthor)
-        private val metadataView: TextView = view.findViewById(R.id.postMetadata)
-        private val bodyView: TextView = view.findViewById(R.id.postBody)
-        private val avatarView: ImageView = view.findViewById(R.id.postAvatar)
-        private val imagesContainer: LinearLayout = view.findViewById(R.id.postImagesContainer)
-        private val commentCountView: TextView = view.findViewById(R.id.postCommentsCount)
-        private val commentsIconView: ImageView = view.findViewById(R.id.postCommentsIcon)
-        private val commentsContainer: View = view.findViewById(R.id.postCommentsContainer)
-        private val editAction: View = view.findViewById(R.id.postActionEdit)
-        private val deleteAction: View = view.findViewById(R.id.postActionDelete)
-        private val shareAction: View = view.findViewById(R.id.postActionShare)
-
-        fun bind(item: DashboardNewsItem) {
-            authorView.text = item.author
-            metadataView.text = item.metadata
-            if (item.message.isNullOrBlank()) {
-                bodyView.isVisible = false
-                bodyView.text = ""
-                bodyView.setOnClickListener(null)
-            } else {
-                bodyView.isVisible = true
-                markwon.setMarkdown(bodyView, item.message)
-                bodyView.setOnClickListener { onPostClicked(item) }
-            }
-            avatarBinder(avatarView, item.username, item.hasAvatar)
-            avatarView.setOnClickListener { onAuthorClicked(item) }
-            authorView.setOnClickListener { onAuthorClicked(item) }
-            metadataView.setOnClickListener { onAuthorClicked(item) }
-            bindImages(item)
-            commentCountView.text = formatCount(item.commentCount)
-            commentsIconView.setOnClickListener { onPostClicked(item) }
-            commentCountView.setOnClickListener { onPostClicked(item) }
-            commentsContainer.setOnClickListener { onPostClicked(item) }
-            bindActions(item)
-        }
-
-        private fun bindActions(item: DashboardNewsItem) {
-            setupAction(editAction, item.canEdit, R.string.dashboard_post_action_edit_toast) {
-                onEditClicked(item)
-            }
-            setupAction(deleteAction, item.canDelete, R.string.dashboard_post_action_delete_toast) {
-                onDeleteClicked(item)
-            }
-            setupAction(shareAction, item.canShare, R.string.dashboard_post_action_share_toast) {
-                onShareClicked(item)
-            }
-        }
-
-        private fun setupAction(
-            view: View,
-            enabled: Boolean,
-            messageRes: Int,
-            onClick: (() -> Unit)? = null,
-        ) {
-            view.isEnabled = enabled
-            view.alpha = if (enabled) 1f else DISABLED_ACTION_ALPHA
-            if (enabled) {
-                view.setOnClickListener { onClick?.invoke() ?: Toast.makeText(view.context, messageRes, Toast.LENGTH_SHORT).show() }
-            } else {
-                view.setOnClickListener(null)
-            }
-        }
-
-        private fun bindImages(item: DashboardNewsItem) {
-            val imagePaths = item.imagePaths
-            if (imagePaths.isEmpty()) {
-                imagesContainer.isVisible = false
-                imagesContainer.removeAllViews()
-                return
-            }
-            imagesContainer.isVisible = true
-            imagesContainer.removeAllViews()
-            val context = imagesContainer.context
-            val spacing = context.resources.getDimensionPixelSize(R.dimen.dashboard_post_image_spacing)
-            val height = context.resources.getDimensionPixelSize(R.dimen.dashboard_card_image_height)
-            imagePaths.forEachIndexed { index, path ->
-                val imageView = AppCompatImageView(context)
-                val params =
-                    LinearLayout.LayoutParams(
-                        android.view.ViewGroup.LayoutParams.MATCH_PARENT,
-                        height,
-                    )
-                if (index > 0) {
-                    params.topMargin = spacing
-                }
-                imageView.layoutParams = params
-                imageView.scaleType = ImageView.ScaleType.CENTER_CROP
-                imageView.setBackgroundResource(R.drawable.dashboard_post_image_placeholder)
-                imageView.contentDescription = context.getString(R.string.dashboard_post_image_content_description)
-                imagesContainer.addView(imageView)
-                imageBinder(imageView, path)
-                imageView.setOnClickListener { onImageClicked(item, index) }
-            }
-        }
-
-        private fun formatCount(count: Int): String {
-            if (count < 0) {
-                return "0"
-            }
-            return COMPACT_DECIMAL_FORMAT.format(count)
-        }
-
-        companion object {
-            private val COMPACT_DECIMAL_FORMAT = CompactDecimalFormat.getInstance(Locale.US, CompactDecimalFormat.CompactStyle.SHORT)
-            private const val DISABLED_ACTION_ALPHA = 0.4f
-        }
-    }
-
     companion object {
         private const val ARG_TEAM_ID = "arg_team_id"
         private const val ARG_TEAM_NAME = "arg_team_name"
-        private const val LOAD_MORE_THRESHOLD = 3
-        private val IMAGE_MARKDOWN_CAPTURE_REGEX = Regex("!\\[[^]]*]\\(([^)]+)\\)")
-        private const val MINUTE_MILLIS = 60_000L
-        private const val HOUR_MILLIS = 60 * MINUTE_MILLIS
-        private const val DAY_MILLIS = 24 * HOUR_MILLIS
-        private const val MONTH_MILLIS = 30 * DAY_MILLIS
-        private const val YEAR_MILLIS = 12 * MONTH_MILLIS
-
         fun newInstanceForTeam(
             teamId: String,
             teamName: String,
